Add optional event order argument to BinaryProvider
The order of events stored in binary files as it was in the original
MDF file is not kept. Since the original ordering is also effectively
undefined, all that is reasonably needed is a way to compare
consistently. This has been implemented by the ability to specify an
optional order argument to the constructor of the BinaryProvider.

In the test that makes use of this, the order is obtained from the
MDFProvider.