Fix repetitions and number of events both being specified
If the number of events (N) is specified together with
repetitions (R), the same N events will be repeated. If N exceeds the
number of available events M, M events will be repeated.

The last slice of a repetition is not filled completely if N or M (as
requested) is not a multiple of the number of events per slice.